SQL repo

+++ src/main/java/com/javaguru/shoppinglist/repository/ProductInMemoryRepository.java
@@ -1,20 +1,22 @@
 package com.javaguru.shoppinglist.repository;
 
 import com.javaguru.shoppinglist.domain.ProductEntity;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 
 import java.util.*;
 
 @Repository
+@Profile("inmemory")
 public class ProductInMemoryRepository implements ProductRepository {
 
     private final Map<Long, ProductEntity> repository = new HashMap<>();
     private Long productIdSequence = 0L;
 
     @Override
-    public void save(ProductEntity product) {
+    public ProductEntity save(ProductEntity product) {
         product.setId(productIdSequence++);
-        repository.put(product.getId(), product);
+        return repository.put(product.getId(), product);
     }
 
     @Override
@@ -23,22 +25,39 @@ public class ProductInMemoryRepository implements ProductRepository {
     }
 
     @Override
-    public List<ProductEntity> findAll() {
-        return new ArrayList<>(repository.values());
+    public Optional<ArrayList<ProductEntity>> findAll() {
+        ArrayList<ProductEntity> enttityList = (ArrayList<ProductEntity>)repository.values();
+        return Optional.of(enttityList);
     }
 
     @Override
-    public Optional<ProductEntity> findByName(String name) {
+    public Boolean isByName(String name) {
         for (ProductEntity entry : repository.values()) {
             if (entry.getName().equals(name)) {
-                return Optional.of(entry);
+                return true;
             }
         }
-        return Optional.empty();
+        return false;
     }
 
     @Override
-    public void deleteByID(Long id) {
-        repository.remove(id);
+    public Boolean deleteByID(Long id) {
+        for (ProductEntity entry : repository.values()) {
+            if (entry.getId().equals(id)) {
+                repository.remove(id);
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public ProductEntity updateById(Long id, ProductEntity entity) {
+        return null;
+    }
+
+    @Override
+    public Boolean isById(Long id) {
+        return null;
     }
 }
